webEnd/requirements.txt
psycopg2==2.9.1
Flask==1.1.2
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
Werkzeug==1.0.1
cryptography==3.4.6
flake8==4.0.1
gunicorn==20.1.0
google-cloud-storage==1.44.0
gcloud==0.18.3
